+++ CHANGES.md
@@ -19,7 +19,7 @@
   - @voluntas
 - [CHANGE] `deps.json` を導入する
   - @voluntas
-- [ADD] text 関連の処理を追加する
+- [ADD] `Pattern` / `Stroke` / `Gradient` / `Text` を追加する
   - @voluntas
 
 ### misc